networking: Fix rlim type format mismatch
- Not all OSes define rlim_cur as the same type (hi there macos), but uint32_t should fit FD count anyhow

+++ src/networking.c
@@ -153,7 +153,7 @@ static void net_init_once()
         if (rlim.rlim_cur < rlim.rlim_max && rlim.rlim_max > 1024) {
             rlim.rlim_cur = rlim.rlim_max;
             if (setrlimit(RLIMIT_NOFILE, &rlim) == 0) {
-                rvvm_info("Raising RLIMIT_NOFILE to %ld", rlim.rlim_cur);
+                rvvm_info("Raising RLIMIT_NOFILE to %u", (uint32_t)rlim.rlim_cur);
             }
         }
     }
